+++ tests/testthat/test-imagecollection.R
@@ -2,28 +2,51 @@
 library(rgee)
 library(raster)
 
-# Initialize a specific Earth Engine account and load
-# either Google Drive or Google Cloud Storage credentials
-ee_path <- path.expand("~/.config/earthengine")
-sessioninfo <- sprintf("%s/rgee_sessioninfo.txt", ee_path)
 
-user <- tryCatch(
-  expr = read.table(sessioninfo,header = TRUE,stringsAsFactors = FALSE),
-  error = function(e) ee_Initialize(
-    email = 'data.colec.fbf@gmail.com',
-    drive = TRUE,
-    gcs = TRUE
-  )
-)
+# Pre-checking ------------------------------------------------------
+# Google credentials were loaded in the system?
+skip_if_no_credentials <- function() {
+  ee_path <- path.expand("~/.config/earthengine")
+  sessioninfo <- sprintf("%s/rgee_sessioninfo.txt", ee_path)
+  if (isFALSE(file.exists(sessioninfo))) {
+    skip("google credentials were not found")
+  }
+}
 
-if (anyNA(user)) {
+# Neccesary Python packages were loaded?
+skip_if_no_pypkg <- function() {
+  have_ee <- reticulate::py_module_available("ee")
+  have_numpy <- reticulate::py_module_available("numpy")
+  if (isFALSE(have_ee)) {
+    skip("ee not available for testing")
+  }
+  if (isFALSE(have_numpy)) {
+    skip("numpy not available for testing")
+  }
+}
+
+# Init Earth Engine just if it is necessary
+init_rgee <- function() {
   ee_reattach()
-  ee_Initialize(
-    email = 'data.colec.fbf@gmail.com',
-    drive = TRUE,
-    gcs = TRUE
+  tryCatch(
+    expr = ee$data$get_persistent_credentials()$client_id,
+    error = function(e) {
+      ee_reattach()
+      ee_Initialize(
+        email = 'data.colec.fbf@gmail.com',
+        drive = TRUE,
+        gcs = TRUE
+      )
+    }
   )
 }
+
+skip_if_no_credentials()
+skip_if_no_pypkg()
+init_rgee()
+
+# -------------------------------------------------------------------------
+
 
 # Dataset
 # ---------------------------