+++ R/zzz.R
@@ -1,7 +1,13 @@
 .onLoad <- function(libname, pkgname) {
-  # use superassignment to update global reference
-  ee <<- reticulate::import("ee", delay_load = FALSE)
+  # Configure a Python environment
   reticulate::configure_environment(pkgname)
+
+  # if EARTHENGINE_PYTHON is defined then forward it to RETICULATE_PYTHON
+  earthengine_python <- Sys.getenv("EARTHENGINE_PYTHON", unset = NA)
+  if (!is.na(earthengine_python))
+    Sys.setenv(RETICULATE_PYTHON = earthengine_python)
+
+  ee <<- reticulate::import("ee", delay_load = TRUE)
 }
 
 .onAttach <- function(libname, pkgname) {
@@ -12,5 +18,5 @@
 #' Reattach ee as a reserved word
 #' @export
 ee_reattach <- function() {
-  ee <<- reticulate::import("ee", delay_load = TRUE)
+  ee <<- reticulate::import("ee", delay_load = FALSE)
 }