+++ NEWS.md
@@ -11,6 +11,10 @@ vignette: >
   %\VignetteIndexEntry{NEWS}
   %\VignetteEncoding{UTF-8}
 ---
+
+# rgee 0.5.1
+- ee_install_* functions were deprecated and replaced by ee_install. ee_install create an isolated Python virtual environment with all rgee dependencies. ee_install can also be used to set the EARTHENGINE_PYTHON variable.
+
 # rgee 0.5.0
 - **Initial CRAN release**
 - GitHub actions supports Windows, Linux and MacOS